disassemble: moving the main logic into a folder
