+++ strcs/meta.py
@@ -7,7 +7,7 @@ import cattrs
 from attrs import define
 
 from . import errors
-from .disassemble import Type, TypeCache
+from .disassemble.base import Type, TypeCache
 
 T = tp.TypeVar("T")
 U = tp.TypeVar("U")
