+++ tests/disassemble/test_base.py
@@ -14,10 +14,9 @@ from attrs import define
 from attrs import has as attrs_has
 
 import strcs
-from strcs import InstanceCheckMeta, Type, resolve_types
-from strcs.disassemble import (
+from strcs import Field, InstanceCheckMeta, Type, resolve_types
+from strcs.disassemble.base import (
     Default,
-    Field,
     fields_from_attrs,
     fields_from_class,
     fields_from_dataclasses,
